# Optimize flutter tizen app build time