@JSUYA 2025-03-13
Could you please update TC too? (test/general/build_targets/package_test.dart)
https://github.com/flutter-tizen/flutter-tizen/actions/runs/13829836858/job/38691455921?pr=595#step:4:278